+++ apps/web/src/types/comment/comment-service/shorts-comment-service.d.ts
@@ -1,4 +1,4 @@
-import type { ShortsReq, ShortsUuid } from "../../shorts-service";
+import type { ShortsReq, ShortsUuid } from "../../shorts/common";
 import type {
   BaseComment,
   DeleteCommentReq,
