+++ apps/web/src/types/shorts/common.d.ts
